backend/requirements_pinecone.txt
google-genai>=0.3.0
pinecone>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
